# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: DDRI API
  description: 강남구 따릉이 대여소 조회·재배치 지원 웹서비스
  version: 1.0.0

servers:
  - url: http://localhost:8000/v1
    description: 로컬 개발

tags:
  - name: ddri-user
    description: 사용자 페이지 API
  - name: ddri-admin
    description: 관리자 페이지 API
  - name: ddri-stations
    description: 스테이션 마스터 API
  - name: weather
    description: 날씨 API (Open-Meteo)

paths:
  /user/stations/nearby:
    get:
      tags: [ddri-user]
      summary: 근처 대여소 조회
      description: |
        사용자 페이지용. lat, lng 기준 거리순 정렬, 지정 시간대 예측.
      parameters:
        - name: lat
          in: query
          required: true
          schema: { type: number, example: 37.5012 }
        - name: lng
          in: query
          required: true
          schema: { type: number, example: 127.0396 }
        - name: target_datetime
          in: query
          required: true
          schema: { type: string, example: "2026-03-18T18:00:00+09:00" }
        - name: limit
          in: query
          schema: { type: integer, default: 20, minimum: 1, maximum: 50 }
        - name: radius_m
          in: query
          schema: { type: integer, nullable: true }
      responses:
        "200":
          description: 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NearbyStationsResponse"

  /admin/stations/risk:
    get:
      tags: [ddri-admin]
      summary: 재배치 판단 목록 조회
      description: |
        관리자 페이지용. 시간대별 위험도·우선순위.
      parameters:
        - name: base_datetime
          in: query
          required: true
          schema: { type: string, example: "2026-03-18T18:00:00+09:00" }
        - name: urgent_only
          in: query
          schema: { type: boolean, nullable: true }
        - name: district_name
          in: query
          schema: { type: string, example: "역삼동" }
        - name: cluster_code
          in: query
          schema: { type: string, example: "cluster00" }
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [risk_score, reallocation_priority, stock_gap]
            default: risk_score
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskStationsResponse"

  /stations:
    get:
      tags: [ddri-stations]
      summary: 스테이션 마스터 조회
      parameters:
        - name: district_name
          in: query
          schema: { type: string }
        - name: cluster_code
          in: query
          schema: { type: string }
      responses:
        "200":
          description: 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StationsListResponse"

  /weather/direct:
    get:
      tags: [weather]
      summary: 날씨 예보 (7일)
      description: |
        Open-Meteo API 직접 조회. start_date 없으면 오늘 포함 7일치, 있으면 해당 날짜부터 최대 16일.
      parameters:
        - name: lat
          in: query
          required: true
          schema: { type: number, example: 37.5665 }
        - name: lon
          in: query
          required: true
          schema: { type: number, example: 126.978 }
        - name: start_date
          in: query
          schema: { type: string, format: date, example: "2026-03-18" }
      responses:
        "200":
          description: 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeatherForecastResponse"
        "400":
          $ref: "#/components/responses/BadRequest"

  /weather/direct/single:
    get:
      tags: [weather]
      summary: 특정일 날씨
      description: |
        Open-Meteo API에서 특정 날짜의 날씨만 조회. target_date 없으면 오늘.
      parameters:
        - name: lat
          in: query
          required: true
          schema: { type: number, example: 37.5665 }
        - name: lon
          in: query
          required: true
          schema: { type: number, example: 126.978 }
        - name: target_date
          in: query
          schema: { type: string, format: date, example: "2026-03-18" }
      responses:
        "200":
          description: 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeatherSingleResponse"
        "400":
          $ref: "#/components/responses/BadRequest"

components:
  schemas:
    NearbyStationsResponse:
      type: object
      properties:
        target_datetime: { type: string }
        user_location: { type: object, properties: { lat: { type: number }, lng: { type: number } } }
        items:
          type: array
          items:
            $ref: "#/components/schemas/StationNearbyItem"
        exceptions:
          type: array
          items:
            $ref: "#/components/schemas/ExceptionItem"

    StationNearbyItem:
      type: object
      properties:
        station_id: { type: integer }
        station_name: { type: string }
        address: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        distance_m: { type: number }
        current_bike_stock: { type: number }
        predicted_rental_count: { type: number }
        predicted_remaining_bikes: { type: number }
        bike_availability_flag: { type: boolean }
        availability_level: { type: string, enum: [sufficient, normal, low] }
        operational_status: { type: string }

    RiskStationsResponse:
      type: object
      properties:
        base_datetime: { type: string }
        summary:
          type: object
          properties:
            total_count: { type: integer }
            risk_count: { type: integer }
            exception_count: { type: integer }
            avg_risk_score: { type: number }
        items:
          type: array
          items:
            $ref: "#/components/schemas/StationRiskItem"
        exceptions:
          type: array
          items:
            $ref: "#/components/schemas/ExceptionItem"

    StationRiskItem:
      type: object
      properties:
        station_id: { type: integer }
        station_name: { type: string }
        district_name: { type: string }
        cluster_code: { type: string }
        current_bike_stock: { type: number }
        predicted_demand: { type: number }
        stock_gap: { type: number }
        risk_score: { type: number }
        reallocation_priority: { type: integer }
        operational_status: { type: string }

    StationsListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/StationMasterItem"
        total_count: { type: integer }

    StationMasterItem:
      type: object
      properties:
        station_id: { type: integer }
        api_station_id: { type: string }
        station_name: { type: string }
        district_name: { type: string }
        address: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        cluster_code: { type: string }
        operational_status: { type: string }

    ExceptionItem:
      type: object
      properties:
        station_id: { type: integer }
        reason: { type: string }

    WeatherForecastResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/WeatherDayItem"
      example:
        results:
          - weather_datetime: "2026-03-18T00:00:00+09:00"
            weather_type: "맑음"
            weather_low: 5.2
            weather_high: 14.1
            icon_url: "https://open-meteo.com/images/weather/..."

    WeatherSingleResponse:
      type: object
      properties:
        result:
          $ref: "#/components/schemas/WeatherDayItem"
      example:
        result:
          weather_datetime: "2026-03-18T00:00:00+09:00"
          weather_type: "맑음"
          weather_low: 5.2
          weather_high: 14.1
          icon_url: "https://open-meteo.com/images/weather/..."

    WeatherDayItem:
      type: object
      properties:
        weather_datetime: { type: string, description: "ISO 8601 날짜·시간" }
        weather_type: { type: string, description: "한글 날씨 유형 (맑음, 흐림 등)" }
        weather_low: { type: number, description: "최저 기온 (°C)" }
        weather_high: { type: number, description: "최고 기온 (°C)" }
        icon_url: { type: string, description: "Open-Meteo 아이콘 URL" }

    ErrorResponse:
      type: object
      properties:
        result: { type: string, example: "Error" }
        errorMsg: { type: string }
        traceback: { type: string, description: "개발 환경에서만" }

  responses:
    BadRequest:
      description: 파라미터 오류
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"